+++ mitogen/sudo.py
@@ -105,7 +105,8 @@ PASSWORD_PROMPT_RE = re.compile(
         r'''
         (?:%s)  # Localised "Password", e.g. "Password", "Mot de passe"
         [^:]*?  # Optional localised text, e.g. "", " for alice", " de alice"
-        :\ ?
+        :
+        (?:\ |\xc2\xa0)?  # Optional SPACE or UTF-8 encoded NO-BREAK SPACE
         \Z  # End of string, prevents repeat matches when pwfeedback echoes '*'
         ''',
     )